[feature/CRUD] add [feature/searchLocation]
