+++ src/pages/MainPages.jsx
@@ -3,10 +3,12 @@ import CompletedDatingCourse from 'components/mainPages/CompletedDatingCourse';
 import MapSearch from 'components/mainPages/MapSearch';
 import SearchResult from 'components/mainPages/SearchResult';
 import React from 'react';
+import Map from 'components/mainPages/Map';
 
 function MainPages() {
   return (
     <>
+      <Map />
       <MapSearch />
       <SearchResult />
       <CompletedDatingCourse />
